Cleanup unused module files and added files_info paramter to prepare new init

+++ poolseq/pipeline/pipeline.py
@@ -13,28 +13,49 @@ class Pipeline():
         self.parameters = Parameters(self.data)
         self.processing = Processing(self.data)
         self.run_list = {'init': self.init, 'clean': self.clean, 'restart': self.restart}
+        self.steps = ('index', 'map', 'sort', 'groups', 'merge', 'duplicates', 'mpileup', 'sync')
         self.run_list[self.parser.arguments.command]()
 
     def init(self):
         self.processing.generate_shell_files(self.data, self.parameters)
 
     def clean(self):
-        results_files = (os.path.join(self.data.directories.results, f) for
-                         f in os.listdir(self.data.directories.results) if
-                         os.path.isfile(os.path.join(self.data.directories.results, f)) or
-                         os.path.isdir(os.path.join(self.data.directories.results, f)))
-        shell_files = (os.path.join(self.data.directories.shell, f) for
-                       f in os.listdir(self.data.directories.shell) if
-                       os.path.isfile(os.path.join(self.data.directories.shell, f)))
-        qsub_files = (os.path.join(self.data.directories.qsub, f) for
-                      f in os.listdir(self.data.directories.qsub) if
-                      os.path.isfile(os.path.join(self.data.directories.qsub, f)))
-        for file in results_files:
-            os.remove(file)
-        for file in shell_files:
-            os.remove(file)
-        for file in qsub_files:
+        if not self.parser.arguments.step:
+            step = 0
+        else:
+            step = self.steps.index(self.parser.arguments.step)
+        print(step)
+        if step < 1:
+            files = [os.path.join(self.data.directories.genomes, f) for
+                     f in os.listdir(self.data.directories.genomes)]
+            for file in files:
+                if file != self.data.genome_path:
+                    os.remove(file)
+        if step < 2:
+            print(self.processing.bwa.mapping.output_file_path)
+            for file in self.processing.bwa.mapping.output_file_path:
+                os.remove(file)
+        if step < 3:
+            for file in self.processing.picard.sort.output_file_path:
+                os.remove(file)
+        if step < 4:
+            for file in self.processing.picard.add_read_groups.output_file_path:
+                os.remove(file)
+        if step < 5:
+            for file in self.processing.picard.merge.output_file_path:
+                os.remove(file)
+        if step < 6:
+            for file in self.processing.picard.mark_duplicates.output_file_path:
+                os.remove(file)
+        if step < 7:
+            for file in self.processing.samtools.mpileup.output_file_path:
+                os.remove(file)
+        for file in self.processing.samtools.mpileup.output_file_path:
             os.remove(file)
 
     def restart(self):
-        pass
+        if not self.parser.arguments.step:
+            pass
+        else:
+            step_n = self.steps.index(self.parser.arguments.step)
+        self.processing.generate_shell_files(self.data, self.parameters, step=step_n)
